Pdf.Content.FontInfo

Description

Font info contains information, extracted from font, that may be needed when processing content stream

Documentation

data FontInfo

Font info

Constructors

FontInfoSimple FISimple

FontInfoComposite FIComposite

data FISimple

Font info for simple fonts

Constructors

FISimple

Fields

fiSimpleUnicodeCMap :: Maybe UnicodeCMap

fiSimpleEncoding :: Maybe SimpleFontEncoding

fiSimpleWidths :: Maybe (Int, Int, [Double])

FirstChar, LastChar, list of widths

fiSimpleFontMatrix :: Transform Double

data FontBaseEncoding

Standard encoding, other encodings are based on them

Constructors

FontBaseEncodingWinAnsi

FontBaseEncodingMacRoman

data SimpleFontEncoding

Encoding fo simple font

Constructors

SimpleFontEncoding

Fields

simpleFontBaseEncoding :: FontBaseEncoding

simpleFontDifferences :: [(Word8, ByteString)]

Mapping from glyph code to glyph name for cases when it is different from base encoding

data FIComposite

Font info for Type0 font

Constructors

FIComposite

Fields

fiCompositeUnicodeCMap :: Maybe UnicodeCMap

fiCompositeWidths :: CIDFontWidths

fiCompositeDefaultWidth :: Double

data CIDFontWidths

Glyph widths for CID fonts

Constructors

CIDFontWidths

Fields

cidFontWidthsChars :: Map Int Double

cidFontWidthsRanges :: [(Int, Int, Double)]

makeCIDFontWidths :: Array -> Either String CIDFontWidths

Make CIDFontWidths from value of "W" key in descendant font

cidFontGetWidth :: CIDFontWidths -> Int -> Maybe Double

Get glyph width by glyph code

fontInfoDecodeGlyphs :: FontInfo -> ByteString -> [(Glyph, Double)]

Decode string into list of glyphs and their widths